[REFACTOR]: 이메일 서비스 infra 패키지로 이동, 인터페이스로 분리
