+++ src/main/java/project/infra/mail/application/EmailService.java
@@ -1,0 +1,6 @@
+package project.infra.mail.application;
+
+public interface EmailService {
+
+    void sendValidateMail(String email, int code);
+}
